"Ui v12 added follow text effect and fuel and red label coloring when empty, need to fix hitbox warping"

+++ core/src/com/kredatus/flockblockers/GameWorld/GameHandler.java
@@ -130,7 +130,7 @@ public class GameHandler implements Screen {
 
         prefs.putInteger("health",airship.health);
         prefs.putInteger("armor",airship.armor);
-        prefs.putInteger("fuel",airship.fuel);
+        prefs.putInteger("fuel",(int)airship.fuel);
         prefs.putInteger("ammo",airship.ammo);
 
 
@@ -176,7 +176,7 @@ public class GameHandler implements Screen {
 
         prefs.putInteger("health",airship.health);
         prefs.putInteger("armor",airship.armor);
-        prefs.putInteger("fuel",airship.fuel);
+        prefs.putInteger("fuel",(int)airship.fuel);
         prefs.putInteger("ammo",airship.ammo);
